feat: merged props parsers into parsePropsToTransformationString
